extract static properties
